api for search + fixes

+++ backend/db/migrations/20240630095846_songs_rating_table.ts
@@ -5,12 +5,12 @@ export async function up(knex: Knex): Promise<void> {
     .createTable("songs_rating", function (table) {
       table.integer("song_id").primary().references("songs.id");
       table.integer("reviews_count").notNullable().defaultTo(0);
-      table.float("rating", 4, 1).notNullable().checkBetween([0, 100]);
-      table.float("relevance", 3, 1).notNullable().checkBetween([0, 10]);
-      table.float("structure", 3, 1).notNullable().checkBetween([0, 10]);
-      table.float("realization", 3, 1).notNullable().checkBetween([0, 10]);
-      table.float("lyrics", 3, 1).notNullable().checkBetween([0, 10]);
-      table.float("beat", 3, 1).notNullable().checkBetween([0, 10]);
+      table.decimal("rating", 4, 1).notNullable().checkBetween([0, 100]);
+      table.decimal("relevance", 3, 1).notNullable().checkBetween([0, 10]);
+      table.decimal("structure", 3, 1).notNullable().checkBetween([0, 10]);
+      table.decimal("realization", 3, 1).notNullable().checkBetween([0, 10]);
+      table.decimal("lyrics", 3, 1).notNullable().checkBetween([0, 10]);
+      table.decimal("beat", 3, 1).notNullable().checkBetween([0, 10]);
     })
     .then(() => 
       knex.schema.raw(`
